pyproject.toml
[tool.poetry]
name = "mnist-sandbox"
version = "0.1.0"
description = "MNIST sandbox for MLops MSU course 2023"
authors = ["rw404 <trapeznikovmisha@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"
pytorch-lightning = "^2.0.9"
tqdm = "^4.66.1"
numpy = "1.24.3"
pandas = "2.0.2"
pathlib = "^1.0.1"
dvc = "^3.30.3"
dvc-gdrive = "^2.20.0"
hydra-core = "^1.3.2"
mlflow = "^2.8.1"
onnx = "^1.15.0"
onnxruntime = "^1.16.3"
matplotlib = "^3.8.2"
torchvision = "^0.16.2"
thop = "^0.1.1.post2209072238"
tritonclient = "^2.41.0"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-design = "^0.5.0"
sphinx-gallery = "^0.14.0"
numpydoc = "^1.6.0"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
